changes to engine

+++ Alternative banking system/engine/src/database/client/PaymentNotification.java
@@ -4,7 +4,7 @@ public class PaymentNotification {
 
     private String loanID;
     private int paymentYaz;
-    private int sumOfPayment;
+    private double sumOfPayment;
 
     public PaymentNotification(String loanID, int paymentYaz, int sumOfPayment) {
         this.loanID = loanID;
@@ -14,5 +14,5 @@ public class PaymentNotification {
 
     public String getLoanID() {return loanID;}
     public int getPaymentYaz() {return paymentYaz;}
-    public int getSumOfPayment() {return sumOfPayment;}
+    public double getSumOfPayment() {return sumOfPayment;}
 }
